Move SVG parsing to SDK
