+++ packages/snaps-sdk/package.json
@@ -39,7 +39,7 @@
     "@metamask/providers": "^14.0.2",
     "@metamask/rpc-errors": "^6.1.0",
     "@metamask/utils": "^8.3.0",
-    "is-svg": "^4.4.0",
+    "fast-xml-parser": "^4.3.4",
     "superstruct": "^1.0.3"
   },
   "devDependencies": {
