+++ packages/snaps-sdk/src/index.ts
@@ -9,6 +9,8 @@ export {
   literal,
   union,
   enumValue,
+  parseSvg,
+  isSvg,
 } from './internals';
 
 // Re-exported from `@metamask/utils` for convenience.
