+++ packages/snaps-sdk/src/internals/index.ts
@@ -2,3 +2,4 @@ export * from './error-wrappers';
 export * from './errors';
 export * from './helpers';
 export * from './structs';
+export * from './svg';
