+++ packages/snaps-utils/src/icon.ts
@@ -1,5 +1,5 @@
+import { isSvg, parseSvg } from '@metamask/snaps-sdk';
 import { assert } from '@metamask/utils';
-import { XMLParser } from 'fast-xml-parser';
 
 import type { VirtualFile } from './virtual-file';
 
@@ -59,45 +59,4 @@ export function getSvgDimensions(svg: string): {
   }
 
   return null;
-}
-
-/**
- * Parse and validate a string as an SVG.
- *
- * @param svg - An SVG string.
- * @returns The SVG, its attributes and children in an object format.
- */
-export function parseSvg(svg: string) {
-  try {
-    const trimmed = svg.trim();
-
-    assert(trimmed.length > 0);
-
-    const parser = new XMLParser({
-      ignoreAttributes: false,
-      parseAttributeValue: true,
-    });
-    const parsed = parser.parse(trimmed, true);
-
-    assert(parsed.svg);
-
-    return parsed.svg;
-  } catch {
-    throw new Error('Snap icon must be a valid SVG.');
-  }
-}
-
-/**
- * Validate that a string is a valid SVG.
- *
- * @param svg - An SVG string.
- * @returns True if the SVG is valid otherwise false.
- */
-export function isSvg(svg: string) {
-  try {
-    parseSvg(svg);
-    return true;
-  } catch {
-    return false;
-  }
 }
